% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places.R
\name{translate_australian}
\alias{translate_australian}
\title{Australian translator}
\usage{
translate_australian(text)
}
\arguments{
\item{text}{Text to be translated.}
}
\description{
Translate from English to Australian accent.
}
\examples{
\dontrun{
translate_australian("What nonsense is this?")
}

}
