% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places.R
\name{translate_austrian}
\alias{translate_austrian}
\title{Austrian translator}
\usage{
translate_austrian(text)
}
\arguments{
\item{text}{Text to be translated.}
}
\description{
Translate from English to Austrian.
}
\examples{
\dontrun{
translate_austrian("What nonsense is this?")
}

}
