% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places.R
\name{translate_boston}
\alias{translate_boston}
\title{Boston translator}
\usage{
translate_boston(text)
}
\arguments{
\item{text}{Text to be translated.}
}
\description{
Translate from English to Boston slang.
}
\examples{
\dontrun{
translate_boston("What nonsense is this?")
}

}
