% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places.R
\name{translate_brooklyn}
\alias{translate_brooklyn}
\title{Brooklyn translator}
\usage{
translate_brooklyn(text)
}
\arguments{
\item{text}{Text to be translated.}
}
\description{
Translate from English to Brooklyn accent.
}
\examples{
\dontrun{
translate_brooklyn("What nonsense is this?")
}

}
