% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character-talks.R
\name{translate_chef}
\alias{translate_chef}
\title{Swedish chef translation}
\usage{
translate_chef(text)
}
\arguments{
\item{text}{Text to be converted.}
}
\description{
Convert from English to Swedish chef speak.
}
\examples{
\dontrun{
translate_chef("What nonsense is this?")
}

}
