% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star-wars.R
\name{translate_cheunh}
\alias{translate_cheunh}
\title{Cheunh translation}
\usage{
translate_cheunh(text)
}
\arguments{
\item{text}{Text to be converted.}
}
\description{
Convert from English to Cheunh language.
}
\examples{
\dontrun{
translate_cheunh("What nonsense is this?")
}

}
