% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places.R
\name{translate_cockney}
\alias{translate_cockney}
\title{Cockney translator}
\usage{
translate_cockney(text)
}
\arguments{
\item{text}{Text to be translated.}
}
\description{
Translate from English to Londoner accent.
}
\examples{
\dontrun{
translate_cockney("What nonsense is this?")
}

}
