% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character-talks.R
\name{translate_dolan}
\alias{translate_dolan}
\title{Dolan translator}
\usage{
translate_dolan(text)
}
\arguments{
\item{text}{Text to be converted.}
}
\description{
Convert from English to Dolan.
}
\examples{
\dontrun{
translate_dolan("What nonsense is this?")
}

}
