% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/got.R
\name{translate_dothraki}
\alias{translate_dothraki}
\title{Dothraki translator}
\usage{
translate_dothraki(text)
}
\arguments{
\item{text}{Text to be converted.}
}
\description{
Convert from English to Dothraki.
}
\examples{
\dontrun{
translate_dothraki("What nonsense is this?")
}

}
