% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nostalgic.R
\name{translate_ebonics}
\alias{translate_ebonics}
\title{Ebonics translator}
\usage{
translate_ebonics(text)
}
\arguments{
\item{text}{Text to be translated.}
}
\description{
Translate from English to ebonics.
}
\examples{
\dontrun{
translate_ebonics("What nonsense is this?")
}

}
