% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internet-fads.R
\name{translate_emoji}
\alias{translate_emoji}
\title{Emoji converter}
\usage{
translate_emoji(text)
}
\arguments{
\item{text}{Text to be converted.}
}
\description{
Convert from text to text filled with relevant emojis.
}
\examples{
\dontrun{
translate_emoji("What nonsense is this?")
}

}
