% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character-talks.R
\name{translate_ferblatin}
\alias{translate_ferblatin}
\title{Ferb Latin translator}
\usage{
translate_ferblatin(text)
}
\arguments{
\item{text}{Text to be converted.}
}
\description{
Convert from English to Ferb Latin.
}
\examples{
\dontrun{
translate_ferblatin("What nonsense is this?")
}

}
