% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character-talks.R
\name{translate_fudd}
\alias{translate_fudd}
\title{Fudd translator}
\usage{
translate_fudd(text)
}
\arguments{
\item{text}{Text to be converted.}
}
\description{
Convert from English to Fudd speak.
}
\examples{
\dontrun{
translate_fudd("What nonsense is this?")
}

}
