% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star-wars.R
\name{translate_gungan}
\alias{translate_gungan}
\title{Gungan translator}
\usage{
translate_gungan(text)
}
\arguments{
\item{text}{Text to be converted.}
}
\description{
Convert from English to Gungan language.
}
\examples{
\dontrun{
translate_gungan("What nonsense is this?")
}

}
