% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star-wars.R
\name{translate_huttese}
\alias{translate_huttese}
\title{Huttese translation}
\usage{
translate_huttese(text)
}
\arguments{
\item{text}{Text to be converted.}
}
\description{
Convert from English to Huttese language.
}
\examples{
\dontrun{
translate_huttese("What nonsense is this?")
}

}
