% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nostalgic.R
\name{translate_jivetalk}
\alias{translate_jivetalk}
\title{Jive talk translator}
\usage{
translate_jivetalk(text)
}
\arguments{
\item{text}{Text to be translated.}
}
\description{
Translate from English to Jive
}
\examples{
\dontrun{
translate_jivetalk("What nonsense is this?")
}

}
