% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star-trek.R
\name{translate_klingon}
\alias{translate_klingon}
\title{Klingon translator}
\usage{
translate_klingon(text)
}
\arguments{
\item{text}{Text to be converted.}
}
\description{
Convert from English to Klingon language.
}
\examples{
\dontrun{
translate_klingon("What nonsense is this?")
}

}
