% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internet-fads.R
\name{translate_leetspeak}
\alias{translate_leetspeak}
\title{Leetspeak converter}
\usage{
translate_leetspeak(text)
}
\arguments{
\item{text}{Text to be converted.}
}
\description{
Convert from English to Leet.
}
\examples{
\dontrun{
translate_leetspeak("What nonsense is this?")
}

}
