% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star-wars.R
\name{translate_mandalorian}
\alias{translate_mandalorian}
\title{Mandalorian translator}
\usage{
translate_mandalorian(text)
}
\arguments{
\item{text}{Text to be converted.}
}
\description{
Convert from English to Mandalorian language.
}
\examples{
\dontrun{
translate_mandalorian("What nonsense is this?")
}

}
