% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character-talks.R
\name{translate_minion}
\alias{translate_minion}
\title{Minion translator}
\usage{
translate_minion(text)
}
\arguments{
\item{text}{Text to be converted.}
}
\description{
Convert from English to Minion speak.
}
\examples{
\dontrun{
translate_minion("What nonsense is this?")
}

}
