% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places.R
\name{translate_norfolk}
\alias{translate_norfolk}
\title{Norfolk translator}
\usage{
translate_norfolk(text)
}
\arguments{
\item{text}{Text to be translated.}
}
\description{
Translate from English to Norfolk dialect.
}
\examples{
\dontrun{
translate_norfolk("What nonsense is this?")
}

}
