% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/english.R
\name{translate_old_english}
\alias{translate_old_english}
\title{Old English translation}
\usage{
translate_old_english(text)
}
\arguments{
\item{text}{Text to be converted.}
}
\description{
Convert from modern English to old English.
}
\examples{
\dontrun{
translate_old_english("What nonsense is this?")
}

}
