% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{translate_piglatin}
\alias{translate_piglatin}
\title{Pig Latin translator}
\usage{
translate_piglatin(text)
}
\arguments{
\item{text}{Text to be translated.}
}
\description{
Translate from English to Pig Latin.
}
\examples{
\dontrun{
translate_piglatin("What nonsense is this?")
}

}
