% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nostalgic.R
\name{translate_pirate}
\alias{translate_pirate}
\title{Pirate translator}
\usage{
translate_pirate(text)
}
\arguments{
\item{text}{Text to be translated.}
}
\description{
Translate from English to Pirate.
}
\examples{
\dontrun{
translate_pirate("What nonsense is this?")
}

}
