% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elvish.R
\name{translate_quenya}
\alias{translate_quenya}
\title{Quenya translator}
\usage{
translate_quenya(text)
}
\arguments{
\item{text}{Text to be converted.}
}
\description{
Convert from English to Quenya.
}
\examples{
\dontrun{
translate_quenya("What nonsense is this?")
}

}
