% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star-trek.R
\name{translate_romulan}
\alias{translate_romulan}
\title{Romulan translation}
\usage{
translate_romulan(text)
}
\arguments{
\item{text}{Text to be converted.}
}
\description{
Convert from English to Romulan language.
}
\examples{
\dontrun{
translate_romulan("What nonsense is this?")
}

}
