% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/english.R
\name{translate_shakespeare}
\alias{translate_shakespeare}
\title{Shakespeare translator}
\usage{
translate_shakespeare(text)
}
\arguments{
\item{text}{Text to be converted.}
}
\description{
Convert from English to Shakespeare
}
\examples{
\dontrun{
translate_shakespeare("What nonsense is this?")
}

}
