% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elvish.R
\name{translate_sindarin}
\alias{translate_sindarin}
\title{Sindarin translator}
\usage{
translate_sindarin(text)
}
\arguments{
\item{text}{Text to be converted.}
}
\description{
Convert from English to Sindarin.
}
\examples{
\dontrun{
translate_sindarin("What nonsense is this?")
}

}
