% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star-wars.R
\name{translate_sith}
\alias{translate_sith}
\title{Sith translation}
\usage{
translate_sith(text)
}
\arguments{
\item{text}{Text to be converted.}
}
\description{
Convert from English to Sith language.
}
\examples{
\dontrun{
translate_sith("What nonsense is this?")
}

}
