% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/english.R
\name{translate_uk_to_us_english}
\alias{translate_uk_to_us_english}
\title{UK to US translator}
\usage{
translate_uk_to_us_english(text)
}
\arguments{
\item{text}{Text to be converted.}
}
\description{
Convert from UK English to US English spelling.
}
\examples{
\dontrun{
translate_uk_to_us_english("What nonsense is this?")
}

}
