% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/english.R
\name{translate_us_to_uk_english}
\alias{translate_us_to_uk_english}
\title{US to UK translator}
\usage{
translate_us_to_uk_english(text)
}
\arguments{
\item{text}{Text to be converted.}
}
\description{
Convert from US English to UK English spelling.
}
\examples{
\dontrun{
translate_us_to_uk_english("What nonsense is this?")
}

}
