% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nostalgic.R
\name{translate_valspeak}
\alias{translate_valspeak}
\title{Valspeak translator}
\usage{
translate_valspeak(text)
}
\arguments{
\item{text}{Text to be translated.}
}
\description{
Translate from English to valspeak.
}
\examples{
\dontrun{
translate_valspeak("What nonsense is this?")
}

}
