% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/got.R
\name{translate_valyrian}
\alias{translate_valyrian}
\title{Valyrian translation}
\usage{
translate_valyrian(text)
}
\arguments{
\item{text}{Text to be converted.}
}
\description{
Convert from English to Valyrian.
}
\examples{
\dontrun{
translate_valyrian("What nonsense is this?")
}

}
