% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star-trek.R
\name{translate_vulcan}
\alias{translate_vulcan}
\title{Vulcan translator}
\usage{
translate_vulcan(text)
}
\arguments{
\item{text}{Text to be converted.}
}
\description{
Convert from English to Vulcan language.
}
\examples{
\dontrun{
translate_vulcan("What nonsense is this?")
}

}
