% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internet-fads.R
\name{translate_wowslang}
\alias{translate_wowslang}
\title{Wowslang	 converter}
\usage{
translate_wowslang(text)
}
\arguments{
\item{text}{Text to be converted.}
}
\description{
Convert from English to World of Warcraft slang.
}
\examples{
\dontrun{
translate_wowslang("What nonsense is this?")
}

}
