% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star-wars.R
\name{translate_yoda}
\alias{translate_yoda}
\title{Yoda translation}
\usage{
translate_yoda(text)
}
\arguments{
\item{text}{Text to be converted.}
}
\description{
Convert from English to Yoda speak.
}
\examples{
\dontrun{
translate_yoda("What nonsense is this?")
}

}
